- Supports PHIZ-35. - Implemented (initial) contextual SSL event logging. - Implemented Spring CGLIB based proxying of SSLContext(s), SSLEngine(s), and (X509Extended)TrustManager(s) in order to capture events. - Implemented automatic existing/valid SSL session invalidation in order to always force (re-)handshaking.

+++ phiz-web-ws/src/main/java/gov/hhs/onc/phiz/web/ws/logging/WsMessageEvent.java
@@ -23,9 +23,9 @@ public interface WsMessageEvent {
     public void setEndpointType(PhizWsEndpointType endpointType);
 
     @JsonProperty
-    public int getEventId();
+    public long getEventId();
 
-    public void setEventId(int eventId);
+    public void setEventId(long eventId);
 
     public Logger getLogger();
